Longitudinal Student Assessment Data

A dataset comprising a subset of five years of state assessment reading data suitable for student growth percentile and percentile growth projection/trajectory analyses. The dataset is used in examples provided in the documentation for the studentGrowthPercentiles and studentGrowthProjections functions, and serves as an exemplar for user construction of their own datasets for growth percentile analyses.

data(sgpData)

Format

A data frame of student level observations for five years on the following variables.

ID

Student ID

GRADE_2013

Student Grade Tested 2013, possibly missing

GRADE_2014

Student Grade Tested 2014, possibly missing

GRADE_2015

Student Grade Tested 2015, possibly missing

GRADE_2016

Student Grade Tested 2016, possibly missing

GRADE_2017

Student Grade Tested 2017, possibly missing

SS_2013

Student Scale Score 2013, possibly missing

SS_2014

Student Scale Score 2014, possibly missing

SS_2015

Student Scale Score 2015, possibly missing

SS_2016

Student Scale Score 2016, possibly missing

SS_2017

Student Scale Score 2017, possibly missing

Note

All datasets used with studentGrowthPercentiles and studentGrowthProjections must be specifically formatted as wide format files. The first variable/column is the student ID variable. The next set of columns provide the grade of the student across all the years provided in the dataset (possibly missing). The last set of columns provide the scales scores of the student in the respective grades. Multi-year operational analyses benefit from putting data in long format. See sgpData_LONG and associated documentation for a comprehensive account.

See also

sgpData_LONG

Source

Anonymized student level state assessment data