Lookup table associating INSTRUCTOR_NUMBER with student ID by YEAR and CONTENT_AREA

Format

A data frame of student level observations for five years and two content areas for the following variables.

ID

Unique Student Identification Number

CONTENT_AREA

Content area for student/teacher record (Reading or Mathematics)

YEAR

Year for student/teacher record

INSTRUCTOR_NUMBER

Unique instructor number identifier

INSTRUCTOR_LAST_NAME

Instructor last name

INSTRUCTOR_FIRST_NAME

Instructor first name

INSTRUCTOR_WEIGHT

Proportional weight associated with student/teacher record

INSTRUCTOR_ENROLLMENT_STATUS

Indicator of full enrollment status of student with teacher

See also

sgpData_LONG

Source

Anonymized student/teacher lookup table in LONG format