Longitudinal Student Assessment Data in LONG format

A dataset comprising a subset of five years of state assessment reading and mathematics data suitable for student growth percentile (SGP) and percentile growth projection/trajectory analyses. The dataset is in LONG format with each record representing a unique student, by year by content area combination. The example data is used in examples and illustration in the enclosed documentation to show how SGPs can be operationalized across years with the results being used to produce a variety of high quality visualizations both at the aggregate (e.g., school) and individual level.

Format

A data frame of student level observations for five years and two content areas for the following variables.

ID

Unique Student Identification Number

LAST_NAME

Student last name

FIRST_NAME

Student first name

CONTENT_AREA

Content area for student observation (Reading or Mathematics)

YEAR

Year for student observation

GRADE

Grade level of student observation

SCALE_SCORE

Student Scale Score

ACHIEVEMENT_LEVEL

Achievement level associated with student scale score

GENDER

Student gender

ETHNICITY

Student ethnicity

FREE_REDUCED_LUNCH_STATUS

Student free/reduced lunch status

ELL_STATUS

Student English Language Learner status

IEP_STATUS

Student Individual Education Plan status

GIFTED_AND_TALENTED_PROGRAM_STATUS

Student Gifted and Talented Program status

SCHOOL_NUMBER

School number associated with student record

SCHOOL_NAME

School name associated with school number and student record

EMH_LEVEL

Elementary, Middle, High school indicator for school attended

DISTRICT_NUMBER

District number associated with student record

DISTRICT_NAME

District name associated with school number and student record

SCHOOL_ENROLLMENT_STATUS

Indicator of full academic year student enrollment in school

DISTRICT_ENROLLMENT_STATUS

Indicator of full academic year student enrollment in school

STATE_ENROLLMENT_STATUS

Indicator of full academic year student enrollment in school

VALID_CASE

Indicator of whether the case is valid or invalid

Note

All datasets used with studentGrowthPercentiles and studentGrowthProjections must be specifically formatted as wide format files. The first variable/column is the student ID variable. The next set of columns provide the grade of the student across all the years provided in the dataset (possibly missing). The last set of columns provide the scales scores of the student in the respective grades. Operational users of this package are advised to use data formatted similar to the sgpData_LONG with the higher level functions prepareSGP, analyzeSGP, combineSGP, summarizeSGP, visualizeSGP, and outputSGP

See also

sgpData

Source

Anonymized student level state assessment data in LONG format